- Added support for mouse wheel resolving issue #37. Not that a particularly clean way to add it really presented itself, but figured by now it might be better to hack it in than to continue not supporting it. - Cleared Clang warnings - Little bit of cleanup of C-isms

+++ src/v_text.cpp
@@ -53,7 +53,7 @@
 //
 // Write a single character using the given font
 //
-void STACK_ARGS DCanvas::DrawChar (FFont *font, int normalcolor, int x, int y, BYTE character, ...)
+void STACK_ARGS DCanvas::DrawChar (FFont *font, int normalcolor, int x, int y, int character, ...)
 {
 	if (font == NULL)
 		return;
